Update session Delete list Cart

+++ resources/views/product/showCart.blade.php
@@ -47,7 +47,7 @@
                             <p class="cart_total_price">{{number_format($item['totalSinglePrice'])}}</p>
                         </td>
                         <td class="cart_delete">
-                            <a class="cart_quantity_delete" href="/product/cart/deleteItemCart/{{$item['data']['id']}}"><i class="fa fa-times"></i></a>
+                            <a class="cart_quantity_delete" href="/product/cart/deleteItemCart/{{$item['data']['id']}}" onclick="return confirm('Are you sure ?')"><i class="fa fa-times"></i></a>
                         </td>
                     </tr>
                     @endforeach
